requirementPricePrediction.txt
numpy
pandas
matplotlib.pyplot
seaborn
base64
pickel
streamlit
sklearn.linear_model import LinearRegression
sklearn.model_selection import train_test_split
sklearn.metrics import r2_score
sklearn.linear_model import LinearRegression
sklearn.tree import DecisionTreeRegressor
sklearn.ensemble import RandomForestRegressor
sklearn.svm import SVR
sklearn.neighbors import KNeighborsRegressor
xgboost import XGBRegressor
sklearn.metrics import mean_absolute_error,mean_squared_error,r2_score